replaced additional descriptions with `See also:`-entry
I removed all descriptions of other functions and inserted an entry in `See also:` instead.
This is related to my first commits' proposal of using a link-description list as `See also:`. 

I also removed the redundant description in `splice!` for removing items. I opted to remove it from the single-item method, as a (negative) range is used to remove items.