Preserve concrete Union types in collect() with generators
When return_type() gives a Union of concrete types, use it as the element
type of the array instead of the type of the first element and progressively
making it broader as needed (often ending with element type Any).

This means that no reallocation will be needed anymore if/when encoutering
an element with a new type. Note that the inferred element type may still be
broader than the actual contents of the array, for example with
(x for x in Union{Int, Void}[1]).

Using a Union element type also has the advantage of using the more efficient
layout for isbitsunion types and of allowing for more efficient generated code
when using the resulting array. This is particularly noticeable with array
comprehensions, which inherit the behavior of collect().