Clarify SubArray indexing behavior
This pull request adds a clarification on the indexing behavior of `SubArray`s. 

I ran into this earlier while reading the existing documentation, which made it sound like indexing a view "out of bounds" of the indices would be allowed so long as those indices were within the bounds of the underlying array. In fact, `getindex` checks bounds when indexing into the view indices, but no bounds check on indexing into the underlying array – those are checked once in `view` (`@boundscheck checkbounds(A, J...)`).