Do not save and restore signal mask when initializing tasks
This is consistent with what we do with `ASM_COPY_STACK`
and it causes new tasks to be initialized to the wrong signal masks now that we initialize
tasks before signal handlers.

This cause ARM core test to fail after #24066 since the task running the test
have SIGINT masked off.