Implement ASM_COPY_STACK on ARM
Also change the AArch64 version to use a relative jump instead of loading the address in register.
This should make the code robust even if the compiler decide to pick a unusual register (fp or lr)
to store the function address.

Add `jl_unreachable` so that the compiler can know that the inline asm never returns.