fix the PyCall build issue, use 16-byte slugs for version dirs
The miniconda installer creates the correct shebang line no problem
but conda update encounters long shebang lines and decides that they
are dangerous, so it silently replaces a "long" (≥128 bytes) path to
`python` with `/usr/bin/env python`. What could go wrong?!? This works
around that behavior by using a 16-byte "slug" in package/version
paths instead of using the full "$uuid/$sha1" string, which, to be
fair, is an excessive 77 characters in the middle of paths.

The issue was that Conda looks for long shebang lines (>127) and
silently replaces the or