Deprecate array of arrays special casing in .= (#24095)
This commit deprecates the special-casing for arrays of arrays when using `A[I...] .=` with entirely scalar indices.  Once this deprecation goes away, `A[1] .= 0` will *always* modify the element that is stored at `A[1]`.  The outer array `A` will not be modified, regardless of its element type.  I have chosen to use the same internal machinery from `@views` for this; when the deprecation is removed the syntactic transform will behave like `broadcast!(identity, @views A[I...], X)`.

fixup!