Explicitly add `-lpthread` when building curl
This is necessary on older Linuces (such as Ubuntu 12.04 x86) that have
older OpenSSL library versions.  Curl switched over to a threaded DNS
resolver by default in version 7.55, and this requires `libpthread`.
However, their build system relies on the fact that OpenSSL (usually)
provides the proper linkage flags, and so breaks on older Linuces.  See
[0] for more details.

[0] https://curl.haxx.se/mail/lib-2017-09/0112.html