Add a multinomial() function to Base
This is roughly as useful as binomial(), so might as well offer it by
default as well. For me personally, it took me a while to realize that
this is how easy it is to implement in terms of binomial(); I'm hoping
that adding it to Base saves others the time it cost me.

One could argue that there's redundancy in the current parameter
specification. E.g. in Sage, one only passes the `k` vector; `n` is
implied from their sum anyway. That makes sense, but I like the
recognizability of the common notation

     /     n     \
    |             |
    | k_1,k_2,k_3 |
     \           /

I'm not married to the idea though and I'd be happy to change it.