Enable FPO again on non-windows or freebsd
This was enabled when we were porting to 3.7 for better unwinding.
The unwinding should have been improved (thanks to registering JIT functions to libunwind
and a new libunwind version) on non-windows a lot that we don't need this anymore.

The performance impact of this is likely small but it makes `code_native` for simple functions
quite a bit simpler and has the biggest yet effect on the code size in system image (-4%).