Don't re-wrap HermOrSym annotated matrices in matrix functions
This now only considers Hermitian/Symmetric as "dispatch wrappers"
that is only used to dispatch to the correct method.
Instead of rewrapping the result in Hermitian/Symmetric we now
return Matrix consistently. Note that this was already the case
for non-positive definite Hermitian{<:Complex} so this change
should help with type stability for some of these functions.

It is also noted that it is very easy, and cost free, to rewrap
the result again if you really need it to be wrapped.