require(mod::Symbol) => require(from::Module, mod::Symbol)
Thread a `from` module through all require-related calls so that we
can look up the name of the module to be loaded in the context of
the module into which it's being loaded. No behavior change yet.

It's a bit unfortunate that `from` is a module and `sym` isn't, but
we can improve argument names later on.