Have ^C throw InterruptException
In general ^C throws InterruptException. The only reason this doesn't
happen in TerminalMenus is because we enable raw mode which makes ^C
just another key sequence. I think it is more coherent to make this
throw an InterruptException as it would in non-raw parts of code.
If the calling code wants to handle this, it can catch the exception.