Exclude version_git.jl from list of dependencies for HTML docs
This file is modified by the light-source-dist target when preparing tarballs.
Depending on it means that the docs will be rebuilt when calling 'make install'
from the tarballs. This requires downloading files, which is problematic
for replicable builds and may not be possible on build servers.

Ref #24496
(cherry picked from commit bd019a4037f59c0ab1db8a08c0d4a4aa23f52ad5)