Fix incorrect use of `exec`/omission of `spawn` in Makefile
In one case `exec` is the wrong thing to use, because it runs the command
and the way it's used here it then tried to run whatever that command
printed on standard output, which is incorrect (we just want to run the
command). In the other case, we needed a spawn, but didn't have it.