Don't mark loads from semi-boxed unions as nonnull
This is a bit of an interesting case. Even though we have concrete
type information for the variable (we know it's an Int on the isa
branch), we're dealing with a semi-unboxed union. We need to make
sure not to claim to know that the pointer is non-null, because
it actually might be (and in fact always is in this example).

Fixes #24632