Fix #24611 (#24694)
Within a Cygwin environment Julia tries to work like a native Windows
program and uses Windows style paths and ENV["USERPROFILE"] for the user
home directory. The credential helper tests which were failing use `git`
found on the PATH. In Cygwin the `git` version expects POSIX style paths
and uses ENV["HOME"] for the user home.

In order to solve this problem I revised the tests to always set the
HOME and USERPROFILE which ensures the user home is always overridden.

Additionally, setting git configuration variables was unnecessary for
these tests so it was removed.