Relocate Circle config and add arch to cache
According to
https://discuss.circleci.com/t/version-isnt-an-allowed-key/12125,
CircleCI may not correctly identify the version in use if the
configuration file is circle.yml in the project's root directory, as
this was the recommended location for Circle v1.0. In v2.0, the
recommended location is .circleci/config.yml. Thus we'll move it over to
that location to ensure that our version is idenified correctly.

To avoid breakages as Circle introduces more architectures, a cache
template key `{{ arch }}` was added. Circle recommends this using this
for any builds that use cached dependencies. Since we use ccache, we'll
add `{{ arch }}` to our cache file names. See
https://discuss.circleci.com/t/use-the-arch-cache-template-key-if-you-rely-on-cached-compiled-binary-dependencies/16129.