Point links in Dates docs to release-0.6 on 0.6
Currently there are links in the Dates section of the manual that point
to files on master. Besides being not generally great, this didn't
really pose a problem until Dates moved to the stdlib on master, which
made these links 404s, thereby causing the docs build to fail on
release-0.6.

The links were adjusted on master in #24821.