Add support for namespaced method call
* Add jl_fieldcall_lookup
* Add fieldcall_lookup wrapper and inference/inlining support
* Lowering support
* Parser and printing support

This patch add preliminary support for calling a function (method in OOB sense)
without having to export the function and therefore solves the namespace
issue for end-users (i.e. users of concrete types).

It doesn't solve the issue for abstract types or implementer of a concrete
extending an abstract type. All the functions that belongs to the
abstract type still need to be imported in order for them to work on the
concrete type.