REPL: adjust the mark when indenting/transposing lines
Right-indenting a line when the mark is at the beginning of the line
was keeping the mark at the beginning, which is odd (corner case),
in particular when the region is visible (active).
The problem was worse when transposing lines down, as the line which was
moved from below to above was becoming part of the region, which clearly
was bug.