faster rand(::SamplerRangeInt)
This sampler was based on the idea that generating a UInt32 value
is cheaper than generating a UInt64 one. Historically, this is
based on the performance characteristics of MersenneTwister.
Here, we refine this idea, knowing that this RNG produces
natively 52 bits of entropy. So the thresholds determining how
native values are generated each time are 52 bits and 104 bits.
In particular, this becomes much more efficient for
small (of length <= 2^52) Int128 ranges, for which previously
3 native values had to be generated instead of one.