lpad, rpad: use character cound instead of textwidth [fix #25016] (#25069)
The general principle here is that we should avoid having behavior in Base that depends on which version of Unicode you're using. It's fine for an external package to provide versions of lpad and rpad which use textwidth and thereby depend on Unicode details, but the functions in Base should not depend on these.