{nextind,prevind}(s, i, 0): index error unless i is valid (#25095)
It's unclear what do when `n == 0` and `i` is not at the start of a character. Should `nextind` return an index before the one you pass? Similarly, for `prevind`, it would be strange to return the same value as for `n == 1`. Accordingly, we throw error in these cases since it's not clear what the right response is.