Update packages.md
It seems that the doc for publishing a new package is outdated, it tells you to 1) register 2) publish and 3) tag, while the correct order is 1) register 2) tag and 3) publish. See: https://github.com/JuliaLang/METADATA.jl/pull/12444

I've reordered the paragraphs accordingly, although since the tagging was a bit long I put the detailed explanations at the end.