intfuncs: short-circuit gcd
When one argument to gcd is equal to 1, the result is fixed
and there is no need to compute the rest.

This makes an enormous difference on random vectors of integers:

    julia> A = BigInt.(rand(Int, 2<<22));

    julia> gcd(A); @time gcd(A)
      0.404063 seconds (16.78 M allocations: 448.000 MiB)
    1

    julia> _gcd(A) = (res=zero(eltype(A)); for a in A; res=gcd(res,a); if res == 1 return res; end; end; return res)
    _gcd (generic function with 1 method)

    julia> _gcd(A); @time _gcd(A)
      0.000028 seconds (38 allocations: 1.086 KiB)
    1

but that's probably not a fair representation of real-world impact, as
no such improvement is to be expected unless we reach 1 very quickly. A
random vector will do that in very few iterations.