Try to handle interrupting runtests better
I had a test that was hanging and I did not know which one. Hitting
^C would kill everything without showing which tests failed and which
tests didn't complete. My next attempt was to enable InterruptException,
but that doesn't seem to play well with the Distributed stuff (a topic
for another time). This works by setting STDIN to raw and monitoring it
for ^C instead.