Mark Symbol as immutable
Symbol is a bit odd. On the one hand it is comparable by identity like mutable
types, on the other hand, its identity is entirely determined by its "contents",
like an immutable type. The setting of the mutabl flag is thus a bit arbitrary.
However, while that is the system perspective, from the user perspective, setting it
to immutable seems clearer, since a user associates the symbol value with the symbol
rather than its identity, and that symbol value is immutable.