Add backtracking to max-sum, improve greedy solver
With this change, the solver is now exact, i.e. it should always
find a solution if it exists, and fail only when it
doesn't, producing a reasonable log trace in that case.
(Optimality is still based on a heuristic, unless the solution
is found by the greedy solver, which is quite improved as well
anyway.)