Clean _build directory before building new docs
Currently non-deploy docs, built with the pretty_urls option disabled,
also get pushed to gh-pages. With this, they will be cleaned away before
the deploy docs are built and we should end up with a single set of HTML
pages on the gh-pages branch (as opposed to the current e.g.
base/file.html AND base/file/index.html).