Allow construction of instantiated Broadcasted{Nothing} objects
This comes up when `flatten`-ing a broadcasted object within a "fallback"
`copyto!` method: `flatten` wants to construct a new Broadcast object and
copy the instantiated information, but we've already destroyed the `Style`
information when we deferred dispatch to the destination! So this simply permits
instantiating `Broadcasted{Nothing}` objects in the sole signature that gets
called by `flatten`.