Add late union splitting for builtins
The case in #25663 presents a problem for codegen, because it contains
a bunch of builtins with union arguments which codegen doesn't like
very much at the moment. This takes the approach of rewriting
`Tuple{Union{A,B}, Int}` to `Union{Tuple{A, Int}, Tuple{B, Int}}`
and union splitting any builtins operating on these late in the
optimization pipeline to make them more amenable to codegen.

The resulting code is extremely ugly, but it does solve the
problem, so good enough for now?

Fixes #25663 (though not the general case where the union contains
non-isbits elements)