optim: enable early de-structuring
If we have an assignment `x = y`, where `x` will be an immutable struct
is only used for certain operations, such that it is valid for `split_struct_alloc`
(notably `getfield`), then we can hoist that getfield operation to the
point of assignment. This may enable further optimization, such as
optimizing or splitting the value of `y`, or freeing GC-references being
held by unused fields.

This operation (getfield) usually occurs lazily in codegen, so hoisting
it should only improve performance, by exposing new optimization opportunities.