Choose element type which can hold replacement in replace()
Choose the element type by promoting the original element type with
that of the replacement(s). This is particularly useful to recode some
values to missing, and it's consistent with what happens e.g. when
concatenating arrays.

The _copy! helper should ideally use Future.copy!, but that function is not
available from Base currently.