Begin work on new optimizer framework
This introduces a new framework for middle-end optimizations in julia, based on
SSA form IR. Much work is still left to be done, but thanks to Jameson's work
this is in a bootstrappable form, so we're merging it behind a feature flag
to allow for quick iteration.

Documentation of the new IR format is available in the devdocs.

Co-authored-by: Jameson Nash <vtjnash@gmail.com>