loading: use realpath of project file for dummy uuid
Otherwise, even if you know which project file you're loading code
with respect to, if there are symlinks and the project file appears
multiple times in the load path, you need to know which of those
load path entries it is being loaded through. Improbably as this may
sound, I actually had this setup on my system, so this is not just
hypothetical.