generate Stdlib registry (#121)
Use `git grep` on each package verison compatible with julia 0.7
to figure out which stdlib packages they use. This may be slightly
wrong in cases of dynamic code loading, but it's a good first-order
approximation of stdlib dependencies. Also generate a registry for
standard library packages so that version resolution can figure out
how to interact with them.