loading: install at packages/$name/$slug instead of no names
Save .ji files at libs/v#.#/$name/$slug.ji so all the caches for a
given package name are in the same directory of that name. These
can be a mix of .ji files for different packages with the same name
but I think that's perfectly acceptable.

Also base the slug code on CRC32 instead of modular arithmetic.