Correct pointer(::AbstractArray, i) to use strides instead of indices
Previously, `pointer` was erroneously calculating memory offsets based upon `linearindices` instead of `strides` and `Base.elsize`.  This corrects it.  This is a bug-fix for StridedArrays that did not have contiguous strides, and arrays that do not specialize `strides` will see the same behavior but with a deprecation warning that the fallback `strides` implementation (that assumes contiguous strides) is deprecated.