build/test: evaluate in module with the right UUID
This is a bit weird since it creates a module inside of the package
module that is being built/tested, called `__build__` or `__test__`,
and then evaluates build and test code inside of that "hidden" sub-
module of the package. This causes the build/test code to find the
correct UUID when it loads its dependencies, allowing that code to
load what the package itself can (and not load what it can't).