Separate AbstractString interface from iteration protocol
Up until now, the basic interface new AbstractStrings had to implement was:

```
struct MyString; ...; end
next(::MyString, i::Int64)::Tuple{Char, Int64}
isvalid(::MyString, i::Int64)::Bool
ncodeunits(::MyString)::Int64
```

In this interface, the iteration state (i.e. the second tuple element returned
from `next`) always had to be the next valid string index. This is inconvenient
for several reasons:
1. The iteration protocol will change, breaking every use of this API
2. Some strings may want iteration states other than linear indicies for efficiency
   reasons (e.g. RopeStrings)
3. Strings implementors can no longer assume that the second argument they receive
   was necessarily produced by them, so may need to do various validation of the
   iteration sate on every iteration.

This PR attempts to remidy this, by introducing a new generic `Next` iterator.
The iterator is defined to iterate (values, next index) tuple, which is the return
value the `next` method on strings at the moment and thus allows for a natural
transition from the older API.
Thus, this PR changes, the requisite method to implement from:

```
next(::MyString, i::Int)::Tuple{Char, Int}
```

to

```
next(::StringNext{MyString}, state::Any)::Tuple{Tuple{Char, Int}, Any}
```

where `StringNext{T} = Next{T, EachIndexString{T}}`

Efficient implementations of iteration over strings, the indicies as well as `Pairs`
can be derived from this iterator. The reason this iterator is useful is perhaps best
understood by considering strings to be variable-length encodings of character arrays.
In a variable-length encoding, one generally decodes the value and the length (i.e. the
index of the next element) at the same time, so it makes sense to base the API
on the implementation of an iterator with these semantics.

To demonstrate the use and test the new abstract implementations based on this iterator,
there are three string types in the test suite:
- CharString, as before, which simply wraps an array of `Chars` with direct indexing. The
  only change to this iterator is to change the signature of the `next` method.
- RopeString, which strings together several Strings, and more importantly does not have
  efficient linear iteration state.
- DecodeString, which decodes escape sequences on the fly as part of iteration. This string
  type demonstrates one string type wrapping another string type to test the interface from
  both sides