Add umlauts to latex_symbols
This adds umlauts (ä, ö, ü, and upper case letters) to the list of latex
symbols recognized by the REPL with tab completion. The LaTeX symbols
are e.g. '\"a' (which matches what *actual* LaTeX uses).

These probably should have been in the W3C symbol mapping file
(unicode.xml), but apparently they are not. This omission is rather
glaring, considering that e.g. \ss => ß and \aa => å *are* defined in
unicode.xml.