Describe three-valued logic in docstrings
Add information regarding 3VL in the presence of missing values to relevant docstrings,
in addition to what already exists in the manual. Also improve docstrings for comparison
operators and make them more consistent with one another. Introduce a new # Implementation
section intended at developers of custom types, and document it in the guidelines.