[NewOptimizer] Add back getfield elim pass
This adds back a (fixed version of) the NewOptimizer getfield elim
pass. This passes all getfield elim (and `@allocated`) tests in base,
with the exception of two tests that check for getfield elim on mutables.
That part is planned in the design of this pass, but will be part of a
follow up PR.