[RTLIB] start implementing soft float16 support.
Some LLVM intrinsics can't be lowered to instructions on target
platforms. Some of these are therefore lowered to libc libcalls
and glibc does not implement all of them. This commit uses `extern_c`
to implement these fallback methods in native Julia.

Similar projects to RTLIB are glibc and compiler-rt.

For now we only implement conversion between `Float16` and `Float32`,
we also cheat by going through `Float32` for the conversion between
`Float64` and `Float16`.