Fix (lin)range with big arguments (#26378)
There are a lot of assumptions in twiceprecision.jl that require a Float16/32/64, but we were hitting those methods in some cases. This adds a fallback and ensures that the _linspace entry points are constrained appropriately. Also use `float(T)` (where `T<:Integer`) to ensure we construct a bigfloat range when appropriate.