WIP: [NewOptimizer] Rewrite inlining pass for new optimizer
Now that the optimizer is complete, we need to start looking at latency.
The biggest problem at the moment is that we keep converting back and
forth to the legacy IR format. In order to cease doing so, we need to
move the other parts of the optimizer that still operate on the old
data structure to operate on the new ones instead. This PR does so
for inlining.

Current status: Boostraps.