Simplify broadcastable a bit further and add docs (#26601)
* Make LinAlg fess up to using Base._return_type

* Simplify broadcastable a bit further and add docs

Amazing what a great exercise writing documentation is. This is a followup to #26435, and more clearly lays out the requirements for defining a custom (non-AbstractArray) type that can participate in broadcasting.  The biggest functional change here is that types that define their own BroadcastStyle must now also implement `broadcastable` to be a no-op.

* Trailing comma, doc fixes from review; more tests