Fix specialized select_value codegen
When inference sees something like:

```
select_value(true, 1::Int, 1.0::Float64)
```

it knows that the condition is constant and can thus
conclude that the result is of `Int` type. Generally,
we would fold this away in inlining. However, if
inlining is disabled for whatever reason. We can end up with

```
select_value(true, 1::Int, 1.0::Float64)::Int
```

in codegen. After the recent enhancements to select_value and
without this patch, this would cause an unconditional trap
to be generated as it tried to convert 1.0 to `Int`. This patch
instead makes sure that the branches of the select have non-empty
intersection with the annotated result type before attempting the
conversion, thus preventing this case from crashing.