Deprecate similar(f, …) in favor of just dispatching directly on f(…) (#26733)
Remove AbstractArray constructors; just use dispatch on functions

Make compiler test use a local function that has the known properties that we want to test. The important thing is not necessarily zeros -- it is a relatively complex function. The `Base.zeros` method tree temporarily breaks this test due to its numerous deprecations, but this will be resolved in the future.