Decouple Broadcasting API from inference
Some of the broadcasting API users still lean on inference -- that can be fixed up later -- but this now no longer hand-feeds them the inferred result of the broadcast.

This has the slight downside that a type-unstable broadcast will not fall back to the simpler `copyto!` method as it must incrementally widen instead. I find this a worthwhile tradeoff.

Also simplify instantiation now that we no longer need to worry about the eltype.