[NewOptimizer] Fix _apply elision (#26821)
The old optimizer had an extra loop outside the inlining pass that would
turn _apply's in to regular calls. When I wrote the new inliner we discussed
that this wasn't actually necessary because we could just keep track of
this information in the inliner (as we do for invoke). However, that of
course also means that if we can't turn something into an :invoke, we
should still at least turn it into a regular call. Do that.