Completely move indexing helpers into wrappers
The key insight here is that these indexing helpers are an _implementation detail_ of an optimization for a particular argument type within a given broadcast implementation. They are not universal across all Broadcasted wrappers -- which is precisely why some styles had wanted to opt out of them. Now the _broadcast_getindex function is solely responsible for allowing indexing into arguments with broadcasted dimensions properly constrained as appropriate. The `Extruded` type pre-computes the dimensions to constrain, allowing an optimization for types who do not statically know this answer -- by default just all `AbstractArray`s.

This still has a performance regression over master in the reduced example `f(r, x) = r .= x.*x.*x.*x` because it does not currently vectorize on this branch. Not sure why.