Make random test less likely to fail
Previously, one of Random's tests was checking `rand(s) != rand(s)`, where `s` was a 11001-length range.  This is expected to have a one-in-11001 chance of failure, and indeed it has caused at least one CI failure. This changes it to something much less likely to spuriously fail -- if I have my probabilities right here I think it'd be on the order of 10^-20.