[GCLowering] Don't consider the ptls getter a safepoint
The fact that this was absent was masking a bug (fixed in a previous commit).
Now, we don't currently allow any safepoints or allocations before the call
to the ptls getter, but in the future we may want to do things like callee-rooted
arguments, which would have to be rooted across the ptls getter if it were a
safepoint. With the bug fixed, we can proactively add this to the list of
known non-safepoints.