[NewOptimizer] Fix codegen test
With the new optimizer enabled, we compile a signature for eval(::Module, ::Any),
which throws off the codegen test. I'm not entirely sure why this is different,
from the old optimizer, but presumably the old optimizer compiled this singature
beforehand. To fix this, simply add another call to `eval` that forces this
signature to be compiled.