Add an easy way to emit assertions in the optimizer
A common situation when working on the compiler is that a pass wants
to replace one value with another that is egal (but constant or otherwise
cheaper to compute). To facilitate debugging such situations, this
commit adds an `:assert_egal` expression that compares its two argument
values and throws an appropriate error if they are not egal. The use case
is simple: The optimizer simply generates both code paths as well as an
inline `:assert_egal`. Then you run the test suite and get a report of
all the places that would have been miscompiles if the optimization had
been active. Another useful thing to do is to dump all IR to a file with
an increasing counter for the label argument in order to match up failing
assertions with the corresponding IR pieces. This is a relatively simple
mechanism, but seems quite useful to debug optimizer problems.